Fix `hazelcast-full-example` config file links (#469)
The links to `hazelcast-full-example` config files go to the maintenance
`z` branch(es) of the `hazelcast` repo which no longer exist.

There's not a direct replacement, but as a workaround we can simply
assume that the `.0` patch version of that releases source code is
publicly accessible and reference that instead.

In `hz-docs` we avoid this issue by having explicit OS/EE versions but
introducing here adds more properties to maintain.