Added detailed documentation on how to start with an Enterprise license (#578)
As part of a POC I am building with the Enterprise version, I need to start Hazelcast and MC with the license.
Since I am using docker and docker compose and repeatedly start/stop the stack, configuring MC license via UI is inconvenient.

When looking for documentation on how to configure the licenses in a compose file, MC's
docs weren't clear enough and I had to click multiple times until I ended up here system-properties.adoc#starting-with-a-license.

This PR adds a section in the getting started guide to expedite the process of setting a license directly in the getting started guide with docker.

---------

Co-authored-by: Amanda Lindsay <v-amanda.lindsay@hazelcast.com>
Co-authored-by: Jack Green <JackPGreen@Gmail.com>